rename $lib/db/settings to $lib/db/keyval, add resetSettings helper
